message dispatching now is optimal

+++ langs/cs/fyrox-lite-sln/fyrox_lite_cs_netcore/src/Auto/NativeClassId.cs
@@ -12,5 +12,5 @@ namespace FyroxLite;
 
 [StructLayout(LayoutKind.Sequential)]
 internal unsafe partial struct NativeClassId {
-    internal int value;
+    internal long value;
 }
